refactor: rename and change interface placement
