+++ src/app/components/NavLink/NavLink.tsx
@@ -1,7 +1,10 @@
-import { Link } from "react-router-dom";
+import { Link, LinkProps } from "react-router-dom";
 import styles from "./navLink.module.scss";
 import React from "react";
-import { NavLinkProps } from "../../types/interfaces";
+
+interface NavLinkProps extends LinkProps {
+    customClassName?: string;
+}
 
 const NavLink = ({customClassName, ...linkProps}: NavLinkProps) => {
     return <Link className={`${styles.link} ${customClassName}`} {...linkProps} />;
