+++ src/app/types/types.ts
@@ -1,0 +1,6 @@
+export {}
+
+export interface JSONGraphRootState {
+    diagramList: [{ cells: [], id: number, diagramName: string }] | null
+    id: number | null
+}
